test(events): restore generic contract ownership
